+++ data-frame/data-frame_test.go
@@ -351,3 +351,31 @@ func TestDataFrame_innerMerge(t *testing.T) {
 	dfb.LoadData(datab)
 	fmt.Println(innerMerge(dfa, dfb, "A"))
 }
+
+func TestDataFrame_Colnames(t *testing.T) {
+	data := [][]string{
+		[]string{"A", "B", "C", "D"},
+		[]string{"1", "2", "3", "4"},
+		[]string{"5", "6", "7", "8"},
+	}
+
+	// Test parsing two columns as integers
+	df := DataFrame{}
+	df.LoadData(data)
+	if fmt.Sprint(df.Names()) != "[A B C D]" {
+		t.Error("Mismatched colnames")
+	}
+
+	err := df.SetNames([]string{})
+	if err == nil {
+		t.Error("Error didn't fired")
+	}
+	err = df.SetNames([]string{"B", "C", "D", "E"})
+	if err != nil {
+		t.Error("Error when setting colnames")
+	}
+
+	if fmt.Sprint(df.Names()) != "[B C D E]" {
+		t.Error("Setter didn't work properly")
+	}
+}