+++ CHANGELOG.md
@@ -2,14 +2,36 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.3.0] - 2016-02-18
+### Added
+- Getter and setter methods for the column names of a DataFrame
+- Bool column type has been made available
+- New Bool() interface
+- A `column` now can now if any of it's elements is NA and a list of
+  said NA elements ([]bool).
+
 ### Changed
-- The `cell` interface has changed. ToInteger() and ToFloat() now
+- Renamed `cell` interface elements to be more idiomatic:
+    - ToInteger() is now Int()
+    - ToFloat() is now Float()
+- The `cell` interface has changed. Int() and Float() now
   return pointers instead of values to prevent future conflicts when
   returning an error. 
 - The `cell` interface has changed. Checksum() [16]byte added.
 - Using cell.Checksum() for identification of unique elements instead
   of raw strings.
+- The `cell` interface has changed, now also requires ToBool() method.
+- String type now does not contain a string, but a pointer to a string.
+
+### Fixed
+- Bool type constructor function Bools now parses `bool` and `[]bool`
+  elements correctly.
+- Int type constructor function Ints now parses `bool` and `[]bool`
+  elements correctly.
+- Float type constructor function Floats now parses `bool` and `[]bool`
+  elements correctly.
+- String type constructor function Strings now parses `bool` and `[]bool`
+  elements correctly.
 
 ## [0.2.1] - 2016-02-14
 ### Fixed
@@ -39,3 +61,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 [0.1.0]:https://github.com/kniren/gota/compare/v0.1.0...v0.1.0
 [0.2.0]:https://github.com/kniren/gota/compare/v0.1.0...v0.2.0
 [0.2.1]:https://github.com/kniren/gota/compare/v0.2.0...v0.2.1
+[0.3.0]:https://github.com/kniren/gota/compare/v0.2.1...v0.3.0